Finalização De Estilização

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,8 +5,11 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 
 use App\Http\Requests\StoreUserRequest;
+use App\Http\Requests\UpdateUserRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -35,7 +38,7 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
-        User::create($request->all());
+        User::create($request->validated());
 
         return redirect()->route('users.index')->with('success', 'Usuário Cadastrado com Sucesso');
     }
@@ -45,7 +48,12 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
-        //
+        if (!$user = User::find($id)) {
+
+            return redirect()->route('users.index')->with('message', 'Usuário não encontrado');
+        }
+
+        return view('admin.users.show', compact(['user']));
     }
 
     /**
@@ -64,14 +72,23 @@ class UserController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(UpdateUserRequest $request, string $id)
     {
         if (!$user = User::find($id)) {
 
             return redirect()->route('users.index')->with('message', 'Usuário não encontrado');
         }
 
-        $user->update($request->all());
+        /*         $data=$request->validated();
+ */
+
+        $data = $request->only('name', 'email');
+   
+        if ($request->password) {
+            $data['password'] = bcrypt($request->password);
+        }
+
+        $user->update($data);
 
         return redirect()->route('users.index')->with('success', 'Usuário Editado com Sucesso');
     }
@@ -81,13 +98,23 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
+       /*  if(Gate::denies('is-admin')){
+
+            return redirect()->route('users.index')->with('message', 'Acesso não autorizado');
+        } */
+
         if (!$user = User::find($id)) {
 
             return redirect()->route('users.index')->with('message', 'Usuário não encontrado');
         }
 
+
+        if($user->id === Auth::user()->id){ 
+            return back()->with('message', 'Você não pode apagar o seu próprio perfil');
+        }
+
         $user->delete();
 
-        return redirect()->route('users.index')->with('success', 'Usuário Editado com Sucesso');
+        return redirect()->route('users.index')->with('success', 'Usuário Apagado com Sucesso');
     }
 }
